moved dirs vendor, images, scripts, pages, blocks to src; updated import addresses; updated readme
