+++ src/blocks/place-card/__description/place-card__description.css
