+++ src/blocks/place-card/__image/place-card__image.css
