+++ src/blocks/popup/__content/_type/popup__content_type_image.css
