--- src/style.css
@@ -1,1 +1,0 @@
-@import url(../pages/index.css);
